Unify IsPoiChecker.
