+++ search/model.hpp
@@ -1,20 +1,12 @@
 #pragma once
 
-#include "indexer/ftypes_matcher.hpp"
-
+#include <cstdint>
 #include <string>
 
 class FeatureType;
 
 namespace search
 {
-/// Describes 2-level POI-exception types that don't belong to any POI-common classes
-/// (amenity, shop, tourism, ...). Used in search algo and search categories index generation.
-class TwoLevelPOIChecker : public ftypes::BaseChecker
-{
-public:
-  TwoLevelPOIChecker();
-};
 
 // This class is used to map feature types to a restricted set of
 // different search classes (do not confuse these classes with search
